requirements.txt
amqp==2.5.2
anyjson==0.3.3
asn1crypto==0.24.0
attrs==19.3.0
backcall==0.1.0
billiard==3.6.3.0
bleach==3.1.5
celery==4.4.2
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.15
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-filter==2.1.0
django-formtools==2.1
django-mssql==1.8
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
entrypoints==0.3
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
httplib2==0.12.3
idna==2.8
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
kombu==4.6.8
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.2
mistune==0.8.4
mysqlclient==1.4.2.post1
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numexpr==2.7.1
numpy==1.14.2
packaging==20.4
pandas==0.20.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.8.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.8.2
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyodbc==4.0.30
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.0
python-dateutil==2.8.1
pytz==2019.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
raven==6.10.0
redis==3.2.1
requests==2.22.0
scikit-learn==0.19.1
scipy==1.4.1
Send2Trash==1.5.0
Serializer==0.2.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tables==3.4.2
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
XlsxWriter==1.1.8
xlwt==1.3.0
zipp==3.1.0
